# openapi.yaml
...
openapi: 3.0.3
info:
  title: Oficina MVP API
  description: Tech Challenge 15SOAT — Sistema de Gestão de Oficina Mecânica (fases
    1 e 2).
  version: 1.0.0
...
paths:
...
  /veiculos:
    get:
      tags:
      - Veículos
      summary: Lista todos os veículos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VeiculoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    post:
      tags:
      - Veículos
      summary: Cadastra um novo veículo
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CadastrarVeiculoRequest"
      responses:
        "201":
          description: Veículo criado
        "400":
          description: Placa ou payload inválido
        "404":
          description: Cliente não encontrado
        "409":
          description: Placa já cadastrada
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /veiculos/cliente/{clienteId}:
    get:
      tags:
      - Veículos
      summary: Lista veículos de um cliente
      parameters:
      - name: clienteId
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VeiculoResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
  /veiculos/{id}:
    get:
      tags:
      - Veículos
      summary: Busca um veículo pelo ID
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "404":
          description: Veículo não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    put:
      tags:
      - Veículos
      summary: Atualiza ficha técnica e/ou dono do veículo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AtualizarVeiculoRequest"
      responses:
        "404":
          description: Veículo ou cliente não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
    delete:
      tags:
      - Veículos
      summary: Remove um veículo
      parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: Removido
        "404":
          description: Veículo não encontrado
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - BearerAuth: []
components:
  schemas:
...
    AtualizarVeiculoRequest:
      required:
      - marca
      - modelo
      type: object
      properties:
        marca:
          description: Marca do veículo
          maxLength: 80
          pattern: \S
          type: string
          example: Toyota
        modelo:
          description: Modelo do veículo
          maxLength: 120
          pattern: \S
          type: string
          example: Corolla
        ano:
          format: int32
          description: Ano de fabricação do veículo
          minimum: 1900
          type: integer
          example: 2020
        clienteId:
          description: ID do cliente proprietário (opcional para atualização)
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174000
...
    CadastrarVeiculoRequest:
      required:
      - placa
      - marca
      - modelo
      - clienteId
      type: object
      properties:
        placa:
          description: Placa do veículo
          maxLength: 10
          pattern: \S
          type: string
          example: ABC1234
        marca:
          description: Marca do veículo
          maxLength: 80
          pattern: \S
          type: string
          example: Toyota
        modelo:
          description: Modelo do veículo
          maxLength: 120
          pattern: \S
          type: string
          example: Corolla
        ano:
          format: int32
          description: Ano de fabricação do veículo
          minimum: 1900
          type: integer
          example: 2020
        clienteId:
          description: ID do cliente proprietário
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
          example: 123e4567-e89b-12d3-a456-426614174000
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...
    VeiculoResponse:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        placa:
          type: string
        marca:
          type: string
        modelo:
          type: string
        ano:
          format: int32
          type: integer
        clienteId:
          $ref: "#/components/schemas/UUID"
  securitySchemes:
    BearerAuth:
      type: http
      description: "JWT emitido pelo POST /auth/login (cole apenas o token, sem prefixo)."
      scheme: bearer
      bearerFormat: JWT
    SecurityScheme:
      type: http
      description: Authentication
      scheme: bearer
      bearerFormat: JWT